+++ src/main/java/com/example/stocksimulation/service/stock/StockService.java
@@ -15,8 +15,6 @@ import java.util.concurrent.ExecutionException;
 @RequiredArgsConstructor
 @Service
 public class StockService {
-    private final String HEARTBEAT_MESSAGE = "HEARTBEAT";
-    private final TaskScheduler taskScheduler = new SimpleAsyncTaskScheduler();
     private final WebSocketHandler handler;
     private WebSocketSession session;
 
@@ -24,15 +22,6 @@ public class StockService {
         StandardWebSocketClient webSocketClient = new StandardWebSocketClient();
         session = webSocketClient.doHandshake(handler, "ws://ops.koreainvestment.com:21000/tryitout/H0STCNT0").get();
         System.out.println("WebSocket client connected to server.");
-
-        taskScheduler.scheduleAtFixedRate(() -> {
-            try {
-                sendHeartbeat();
-                System.out.println("heart-beat");
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-        }, 80000);
     }
 
     public void test() throws IOException, ExecutionException, InterruptedException {
@@ -42,14 +31,6 @@ public class StockService {
     public void send(String message) throws IOException {
         if (session != null && session.isOpen()) {
             session.sendMessage(new TextMessage(message));
-        } else {
-            System.out.println("WebSocket session is not open.");
-        }
-    }
-
-    private void sendHeartbeat() throws IOException {
-        if (session != null && session.isOpen()) {
-            session.sendMessage(new TextMessage(HEARTBEAT_MESSAGE));
         } else {
             System.out.println("WebSocket session is not open.");
         }
